+++ components/homepage/css/ChallengeDashboard.module.css
@@ -1,0 +1,10 @@
+.Challenges {
+    padding: 30px 15px 10px;
+    gap: 8px;
+
+
+    width: 1043px;
+
+    box-shadow: 0px 2px 8px rgba(0, 0, 0, 0.25);
+    border-radius: 10px;
+}
