--- components/homepage/css/Challenges.module.css
@@ -1,6 +1,0 @@
-.Challenges {
-    padding: 30px 0px;
-    gap: 35px;
-
-    width: 1043px;
-}
