+++ components/navbar/css/NavBar.module.css
@@ -34,7 +34,6 @@
     display: flex;
     flex-direction: row;
     align-items: flex-start;
-    padding: 0px 343px 0px 15px;
     gap: 36px;
 }
