+++ lib/navLinks.js
@@ -1,7 +1,7 @@
 export const navLinks = [
     {
         name: "Home",
-        path: "/home"
+        path: "/"
     },
     {
         name: "My Challenge",
