+++ lib/taskList.js
@@ -1,0 +1,20 @@
+export const taskList = [
+    {
+        taskName: 'This is Task name',
+        challengeName: 'This is Challenge name',
+        dueDate: Date(),
+        finished: true
+    } ,
+    {
+        taskName: 'This is Task name',
+        challengeName: 'This is Challenge name',
+        dueDate: Date(),
+        finished: true
+    } ,
+    {
+        taskName: 'This is Task name',
+        challengeName: 'This is Challenge name',
+        dueDate: Date(),
+        finished: true
+    } ,
+]
